Fixed float to int for lightlevel

+++ main.c
@@ -49,7 +49,7 @@ struct sector{
     char textfloor[9];
     char textceil[9]; //texturefloor = "FLOOR0_1"; textureceiling = "CEIL1_1";
     int special;
-    float lightlevel; //Can be interesing if we want to cast fixed shadow from terrain... (For the future heh)
+    int lightlevel; //Can be interesing if we want to cast fixed shadow from terrain... (For the future heh)
     unsigned int id; //Sector ID
 };
 
@@ -167,9 +167,9 @@ int main(int argc, char *argv[]){
         }*/
         //fprintf(fptr,"sector // %d\n{\nheightfloor = %d;\nheightceiling = %d;\ntexturefloor = \"%s\";\ntextureceiling = \"%s\";\nlightlevel = %d;\nspecial = %d;\nid = %d;\n}\n\n",sectors[i].id,sectors[i].floor,sectors[i].ceil,sectors[i].textfloor,sectors[i].textceil,160,0,sectors[i].id);
         if(sectors[i].special==0){
-            fprintf(fptr,"sector // %d\n{\nheightfloor = %d;\nheightceiling = %d;\ntexturefloor = \"%s\";\ntextureceiling = \"%s\";\nlightlevel = %f;\n}\n\n",sectors[i].id,sectors[i].floor,sectors[i].ceil,sectors[i].textfloor,sectors[i].textceil,sectors[i].lightlevel);
+            fprintf(fptr,"sector // %d\n{\nheightfloor = %d;\nheightceiling = %d;\ntexturefloor = \"%s\";\ntextureceiling = \"%s\";\nlightlevel = %d;\n}\n\n",sectors[i].id,sectors[i].floor,sectors[i].ceil,sectors[i].textfloor,sectors[i].textceil,sectors[i].lightlevel);
         }else{
-            fprintf(fptr,"sector // %d\n{\nheightfloor = %d;\nheightceiling = %d;\ntexturefloor = \"%s\";\ntextureceiling = \"%s\";\nlightlevel = %f;\nspecial = %d;\n}\n\n",sectors[i].id,sectors[i].floor,sectors[i].ceil,sectors[i].textfloor,sectors[i].textceil,sectors[i].lightlevel,sectors[i].special);
+            fprintf(fptr,"sector // %d\n{\nheightfloor = %d;\nheightceiling = %d;\ntexturefloor = \"%s\";\ntextureceiling = \"%s\";\nlightlevel = %d;\nspecial = %d;\n}\n\n",sectors[i].id,sectors[i].floor,sectors[i].ceil,sectors[i].textfloor,sectors[i].textceil,sectors[i].lightlevel,sectors[i].special);
         }
         if(C_SIZE>=4){
             if(i%(2*(C_SIZE-1)*(C_SIZE-1)/10)==0 &&i!=0){
@@ -602,7 +602,7 @@ void generateSectors(){
             }
             
             strcpy(sectors[sectorcounter].textceil,"F_SKY1");                
-            sectors[sectorcounter].lightlevel=160.0; 
+            sectors[sectorcounter].lightlevel=160; 
             sectors[sectorcounter].id=sectorcounter;
             sectorcounter++;
             
@@ -617,7 +617,7 @@ void generateSectors(){
                 strcpy(sectors[sectorcounter].textfloor,"GRASS1");
             }
             strcpy(sectors[sectorcounter].textceil,"F_SKY1");                
-            sectors[sectorcounter].lightlevel=160.0; 
+            sectors[sectorcounter].lightlevel=160; 
             sectors[sectorcounter].id=sectorcounter;
             sectorcounter++;
